Fix the error message of tsdata call sequence (#4244)
* fix scaler(fit=False)

* fix call sequence

* fix error message

* fix check fitted

* autopep8

* added gen_global,gen_roll and change check_is_fitted.

* fix pep8

+++ python/chronos/src/bigdl/chronos/data/utils/feature.py
@@ -31,7 +31,7 @@ def _is_busy_hours(hour):
 
 
 def _is_weekend(weekday):
-    return (weekday >= 5).values
+    return (weekday >= 5).values.astype(np.int64)
 
 
 TIME_FEATURE = ("MINUTE", "DAY", "DAYOFYEAR", "HOUR", "WEEKDAY", "WEEKOFYEAR", "MONTH")
